Conflicts:
	training/training-dao/src/main/java/co/com/expertla/training/dao/impl/configuration/ActivityDaoImpl.java
	training/training-model/src/main/java/co/com/expertla/training/model/dto/ActivityDTO.java
	training/training-model/src/main/java/co/com/expertla/training/model/entities/Modality.java
	training/training-service/src/main/java/co/com/expertla/training/service/configuration/ActivityService.java
	training/training-service/src/main/java/co/com/expertla/training/service/impl/configuration/ActivityServiceImpl.java
	training/training-web/src/main/java/co/com/expertla/training/web/controller/configuration/ActivityController.java
	training/training-web/src/main/java/co/com/expertla/training/web/controller/configuration/ModalityController.java
	training/training-web/src/main/webapp/static/js/login/controller/mainController.js

+++ training/training-model/src/main/java/co/com/expertla/training/model/entities/CoachAssignedPlan.java
@@ -41,7 +41,7 @@ public class CoachAssignedPlan implements Serializable {
     @Column(name = "coach_assigned_plan_id")
     private Integer coachAssignedPlanId;
     @Column(name = "state_id")
-    private Integer stateId;
+    private Short stateId;
     @Column(name = "creation_date")
     @Temporal(TemporalType.DATE)
     private Date creationDate;
@@ -67,11 +67,11 @@ public class CoachAssignedPlan implements Serializable {
         this.coachAssignedPlanId = coachAssignedPlanId;
     }
 
-    public Integer getStateId() {
+    public Short getStateId() {
         return stateId;
     }
 
-    public void setStateId(Integer stateId) {
+    public void setStateId(Short stateId) {
         this.stateId = stateId;
     }
